change some db tables and reformat folder format

+++ Loloca-BE/Data/Entities/Order.cs
@@ -10,7 +10,7 @@ namespace Loloca_BE.Data.Entities
         public int? BookingTourRequestsId { get; set; }
         public int? BookingTourGuideRequestId { get; set; }
         public string OrderCode { get; set; } = null!;
-        public double OrderPrice { get; set; }
+        public decimal OrderPrice { get; set; }
         public string? PaymentProvider { get; set; }
         public string? TransactionCode { get; set; }
         public int Status { get; set; }
